+++ src/components/CartProduct/index.js
@@ -1,0 +1,3 @@
+import CartProduct from './CartProduct';
+
+export default CartProduct;
