--- src/pages/Cartproducts/index.js
@@ -1,3 +1,0 @@
-import CartProducts from './CartProducts';
-
-export default CartProducts;
